v10
L'asservissement Angulaire fonctionne

+++ Robot_Minh_Theo.X/UART_Protocol.c
@@ -141,7 +141,7 @@ void UartProcessDecodedMessage(int function, int payloadLength, unsigned char* p
             getBytesFromFloat(payload_PidX, 0, (float) (kp));
             getBytesFromFloat(payload_PidX, 4, (float) (ki));
             getBytesFromFloat(payload_PidX, 8, (float) (kd));
-            SetupPidAsservissement(&robotState.PidX, kp, ki, kd, 100, 100, 100);
+            SetupPidAsservissement(&robotState.PidX, (double)kp, (double)ki, (double)kd, 100, 100, 100);
 
 
             UartEncodeAndSendMessage(SetPIDX, 12, payload_PidX);
@@ -155,7 +155,7 @@ void UartProcessDecodedMessage(int function, int payloadLength, unsigned char* p
             getBytesFromFloat(payload_PidT, 0, (float) (kp));
             getBytesFromFloat(payload_PidT, 4, (float) (ki));
             getBytesFromFloat(payload_PidT, 8, (float) (kd));
-            SetupPidAsservissement(&robotState.PidTheta, kp, ki, kd, 100, 100, 100);
+            SetupPidAsservissement(&robotState.PidTheta, (double)kp,(double) ki, (double)kd, 100, 100, 100);
 
 
             UartEncodeAndSendMessage(SetPIDT, 12, payload_PidT);
